+++ CHANGELOG.md
@@ -23,6 +23,13 @@
 - [`parity-multiaddr` CHANGELOG](misc/multiaddr/CHANGELOG.md)
 - [`libp2p-core-derive` CHANGELOG](misc/core-derive/CHANGELOG.md)
 
+# Version 0.25.0 [unreleased]
+
+- The `ProtocolsHandler` in `libp2p-swarm` has a new associated type
+  `InboundOpenInfo` ([PR 1714]).
+
+[PR 1714]: https://github.com/libp2p/rust-libp2p/pull/1714
+
 # Version 0.24.0 [2020-08-18]
 
 - Update `libp2p-core`, `libp2p-gossipsub`, `libp2p-kad`, `libp2p-mdns`,