+++ protocols/request-response/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.3.0 [unreleased]
+
+- Update `libp2p-swarm`.
+
 # 0.2.0 [2020-08-18]
 
 - Fixed connection keep-alive, permitting connections to close due